+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [0.7.0] - unreleased
+## [0.7.0] - 10/06/2024
 
 ### Added